backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
google-search-results==2.4.2
pydantic==2.5.3
anthropic==0.39.0
